Add API Handling for Lulzbot
I've implemented a separate set of code to handle the Lulzbot printer.

This is a simple copy-paste of the code used for the Type A Print Pod,
with only a few changes meant to handle the different API key, as well
as the different hostname format.

Future improvements will include breaking this program into functions
and reducing this unnecessary code reuse.